# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Archie finance API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
  "servers": [],
  "paths": {
...
    "/v1/user/email-verification": {
      "get": {
        "operationId": "UserController_checkEmailVerification",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEmailVerificationResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/user/email-verification/resend": {
      "post": {
        "operationId": "UserController_resendEmailVerification",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/user/mfa/enroll": {
      "post": {
        "operationId": "UserController_enrollMfa",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSendEnrollmentTicketResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/user/mfa/enrollments/{enrollmentId}": {
      "delete": {
        "operationId": "UserController_deleteMfaEnrollment",
        "parameters": [
          {
            "name": "enrollmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": "ENROLLMENT_NOT_FOUND"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/user/mfa/enrollments": {
      "get": {
        "operationId": "UserController_getMfaEnrollments",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetEnrollmentResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/user/mfa/enrollment": {
      "get": {
        "operationId": "UserController_isMfaEnrolled",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetMfaEnrollmentResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GetEmailVerificationResponse": {
        "type": "object",
        "properties": {
          "isVerified": {
            "type": "boolean"
          },
          "email": {
            "type": "string"
          }
        },
        "required": [
          "isVerified",
          "email"
        ]
      },
      "GetSendEnrollmentTicketResponse": {
        "type": "object",
        "properties": {
          "ticket_id": {
            "type": "string"
          },
          "ticket_url": {
            "type": "string"
          }
        },
        "required": [
          "ticket_id",
          "ticket_url"
        ]
      },
      "GetEnrollmentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "object"
          },
          "enrolled_at": {
            "type": "string"
          },
          "last_auth": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "auth_method": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "status",
          "enrolled_at",
          "last_auth",
          "type",
          "auth_method"
        ]
      },
      "GetMfaEnrollmentResponse": {
        "type": "object",
        "properties": {
          "isEnrolled": {
            "type": "boolean"
          }
        },
        "required": [
          "isEnrolled"
        ]
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    }
  }
}